requirements.txt
fastapi
uvicorn
oracledb
python-jose
numpy